test(KEN-870): the gate's fixtures are a module its suites share [no-changelog]
The pre-commit gate's suite crossed the size ratchet's cap for a test file
on the merge of this branch with main, where neither side crossed it alone.
Test rows never rise, so the file comes down instead: the fixtures, the
environment isolation, the narrowed PATH and the two verdict helpers move to
a sibling module, and the construct test moves to its own suite beside it.
Every case that ran before runs now.
